Front-end organization overhaul part 2
