+++ app/javascript/packs/vue_setup.js
@@ -1,13 +1,37 @@
 import Vue from 'vue/dist/vue.esm'
 
 document.addEventListener('DOMContentLoaded', () => {
+
+  /*
+   * Pages
+   */
+  if (document.querySelector('#admin-conflicts-index')) {
+    new Vue({
+      el: '#admin-conflicts-index',
+      components: {
+        AdminConflictsIndex: require('../vue/pages/admin/conflicts_index.vue').default,
+      },
+    })
+  }
+
+  if (document.querySelector('#admin-payments-index')) {
+    new Vue({
+      el: '#admin-payments-index',
+      components: {
+        AdminPaymentsIndex: require('../vue/pages/admin/payments_index.vue').default,
+      },
+    })
+  }
+
+  /*
+   * Components
+   */
   if (document.querySelector('#omni-bar')) {
     new Vue({
       el: '#omni-bar',
       components: {
         OmniBar: require('../vue/components/omni_bar.vue').default
       },
-      data: {},
     })
   }
 
@@ -17,7 +41,6 @@ document.addEventListener('DOMContentLoaded', () => {
       components: {
         PaymentSchedule: require('../vue/components/payment_schedule.vue').default
       },
-      data: {},
     })
   }
 
@@ -28,7 +51,6 @@ document.addEventListener('DOMContentLoaded', () => {
       components: {
         UserList: require('../vue/components/user_list.vue').default
       },
-      data: {},
     })
   }
 
@@ -38,7 +60,6 @@ document.addEventListener('DOMContentLoaded', () => {
       components: {
         UserList: require('../vue/components/user_list.vue').default
       },
-      data: {},
     })
   }
 })
