+++ app/javascript/vue/components/recent_payments.vue
@@ -51,9 +51,9 @@
 </template>
 
 <script>
-import Utilities from '../packs/utilities'
+import Utilities from '../../packs/utilities'
+import Toast from '../../packs/toast'
 import moment from 'moment/moment'
-import Toast from '../packs/toast'
 
 export default {
   data: () => ({
